requirements.txt
requests~=2.27.1
discord-py-interactions~=4.3.2
python-dotenv~=1.0.0